# Improve decodeParamsLine by not use regex