--- a/reader.go
+++ b/reader.go
@@ -4,7 +4,7 @@ package m3u8
  Part of M3U8 parser & generator library.
  This file defines functions related to playlist parsing.
 
- Copyleft 2013-2014 Alexander I.Grafov aka Axel <grafov@gmail.com>
+ Copyleft 2013-2015 Alexander I.Grafov aka Axel <grafov@gmail.com>
 
  This program is free software: you can redistribute it and/or modify
  it under the terms of the GNU General Public License as published by