--- a/reader.go
+++ b/reader.go
@@ -444,35 +444,51 @@ func decodeLineOfMediaPlaylist(p *MediaPlaylist, wv *WV, state *decodingState, l
 		}
 	case strings.HasPrefix(line, "#EXT-X-KEY:"):
 		state.listType = MEDIA
-		state.key = new(Key)
+		state.xkey = new(Key)
 		for _, param := range strings.Split(line[11:], ",") {
 			if strings.HasPrefix(param, "METHOD=") {
-				if _, err = fmt.Sscanf(param, "METHOD=%s", &state.key.Method); strict && err != nil {
+				if _, err = fmt.Sscanf(param, "METHOD=%s", &state.xkey.Method); strict && err != nil {
 					return err
 				}
 			}
 			if strings.HasPrefix(param, "URI=") {
-				if _, err = fmt.Sscanf(param, "URI=%s", &state.key.URI); strict && err != nil {
+				if _, err = fmt.Sscanf(param, "URI=%s", &state.xkey.URI); strict && err != nil {
 					return err
 				}
 			}
 			if strings.HasPrefix(param, "IV=") {
-				if _, err = fmt.Sscanf(param, "IV=%s", &state.key.IV); strict && err != nil {
+				if _, err = fmt.Sscanf(param, "IV=%s", &state.xkey.IV); strict && err != nil {
 					return err
 				}
 			}
 			if strings.HasPrefix(param, "KEYFORMAT=") {
-				if _, err = fmt.Sscanf(param, "KEYFORMAT=%s", &state.key.Keyformat); strict && err != nil {
+				if _, err = fmt.Sscanf(param, "KEYFORMAT=%s", &state.xkey.Keyformat); strict && err != nil {
 					return err
 				}
 			}
 			if strings.HasPrefix(param, "KEYFORMATVERSIONS=") {
-				if _, err = fmt.Sscanf(param, "KEYFORMATVERSIONS=%s", &state.key.Keyformatversions); strict && err != nil {
+				if _, err = fmt.Sscanf(param, "KEYFORMATVERSIONS=%s", &state.xkey.Keyformatversions); strict && err != nil {
 					return err
 				}
 			}
 		}
 		state.tagKey = true
+	case strings.HasPrefix(line, "#EXT-X-MAP:"):
+		state.listType = MEDIA
+		state.xmap = new(Map)
+		for _, param := range strings.Split(line[11:], ",") {
+			if strings.HasPrefix(param, "URI=") {
+				if _, err = fmt.Sscanf(param, "URI=%s", &state.xmap.URI); strict && err != nil {
+					return err
+				}
+			}
+			if strings.HasPrefix(param, "BYTERANGE=") {
+				if _, err = fmt.Sscanf(param, "BYTERANGE=%d@%d", &state.xmap.Limit, &state.xmap.Offset); strict && err != nil {
+					return fmt.Errorf("Byterange sub-range length value parsing error: %s", err)
+				}
+			}
+		}
+		state.tagMap = true
 	case !state.tagProgramDateTime && strings.HasPrefix(line, "#EXT-X-PROGRAM-DATE-TIME:"):
 		state.tagProgramDateTime = true
 		state.listType = MEDIA
@@ -527,13 +543,23 @@ func decodeLineOfMediaPlaylist(p *MediaPlaylist, wv *WV, state *decodingState, l
 		}
 		// If EXT-X-KEY appeared before reference to segment (EXTINF) then it linked to this segment
 		if state.tagKey {
-			p.Segments[(p.tail-1)%p.capacity].Key = &Key{state.key.Method, state.key.URI, state.key.IV, state.key.Keyformat, state.key.Keyformatversions}
+			p.Segments[(p.tail-1)%p.capacity].Key = &Key{state.xkey.Method, state.xkey.URI, state.xkey.IV, state.xkey.Keyformat, state.xkey.Keyformatversions}
 			// First EXT-X-KEY may appeared in the header of the playlist and linked to first segment
 			// but for convenient playlist generation it also linked as default playlist key
 			if p.Key == nil {
-				p.Key = state.key
+				p.Key = state.xkey
 			}
 			state.tagKey = false
+		}
+		// If EXT-X-MAP appeared before reference to segment (EXTINF) then it linked to this segment
+		if state.tagMap {
+			p.Segments[(p.tail-1)%p.capacity].Map = &Map{state.xmap.URI, state.xmap.Limit, state.xmap.Offset}
+			// First EXT-X-MAP may appeared in the header of the playlist and linked to first segment
+			// but for convenient playlist generation it also linked as default playlist map
+			if p.Map == nil {
+				p.Map = state.xmap
+			}
+			state.tagMap = false
 		}
 	case strings.HasPrefix(line, "#WV-AUDIO-CHANNELS"):
 		state.listType = MEDIA