- Improve decodeParamsLine by not use regex

Before:

```
BenchmarkDecodeMasterPlaylist-8    30000         45812 ns/op       12197 B/op         77 allocs/op
BenchmarkDecodeMediaPlaylist-8       100      17077647 ns/op    11000939 B/op     120102 allocs/op
BenchmarkEncodeMasterPlaylist-8  1000000          1069 ns/op          40 B/op         11 allocs/op
BenchmarkEncodeMediaPlaylist-8       300       4766209 ns/op       45425 B/op        405 allocs/op
```

After:

```
BenchmarkDecodeMasterPlaylist-8    50000         31298 ns/op        9716 B/op         42 allocs/op
BenchmarkDecodeMediaPlaylist-8       100      16862658 ns/op    11000960 B/op     120104 allocs/op
BenchmarkEncodeMasterPlaylist-8  1000000          1068 ns/op          40 B/op         11 allocs/op
BenchmarkEncodeMediaPlaylist-8       300       4726820 ns/op       45426 B/op        405 allocs/op
```

Benchstat Summary:

```
name                    old time/op    new time/op    delta
DecodeMasterPlaylist-8    45.8µs ± 0%    31.3µs ± 0%   ~             (p=1.000 n=1+1)
DecodeMediaPlaylist-8     17.1ms ± 0%    16.9ms ± 0%   ~             (p=1.000 n=1+1)
EncodeMasterPlaylist-8    1.07µs ± 0%    1.07µs ± 0%   ~             (p=1.000 n=1+1)
EncodeMediaPlaylist-8     4.77ms ± 0%    4.73ms ± 0%   ~             (p=1.000 n=1+1)

name                    old alloc/op   new alloc/op   delta
DecodeMasterPlaylist-8    12.2kB ± 0%     9.7kB ± 0%   ~             (p=1.000 n=1+1)
DecodeMediaPlaylist-8     11.0MB ± 0%    11.0MB ± 0%   ~             (p=1.000 n=1+1)
EncodeMasterPlaylist-8     40.0B ± 0%     40.0B ± 0%   ~     (all samples are equal)
EncodeMediaPlaylist-8     45.4kB ± 0%    45.4kB ± 0%   ~             (p=1.000 n=1+1)

name                    old allocs/op  new allocs/op  delta
DecodeMasterPlaylist-8      77.0 ± 0%      42.0 ± 0%   ~             (p=1.000 n=1+1)
DecodeMediaPlaylist-8       120k ± 0%      120k ± 0%   ~             (p=1.000 n=1+1)
EncodeMasterPlaylist-8      11.0 ± 0%      11.0 ± 0%   ~     (all samples are equal)
EncodeMediaPlaylist-8        405 ± 0%       405 ± 0%   ~     (all samples are equal)
```