@bradleyfalzon 2016-07-06
Hi,

I'm not certain whether we want to merge this change, although there is a performance benefit, the additional lines of code (~50) to support it may not be justified.

Does your use case really require this performance improvement?