@hori-ryota 2016-07-06
I run animation service, and there are a lot of programs more than 24 hours.
I promote efficiency using CDN, but am glad for me when high-speed as possible.
However, because I understand your opinion very much, I think to withdraw it if I think that this change is not good.